Proyecto 1
Aquí va la versión 1.0 de Durini. Hace falta modificar el callback
privado.
