+++ Programa1/Programa1/Debug/Programa1.log
@@ -1,0 +1,4 @@
+﻿  Programa1.cpp
+c:\users\jcdur\google drive\landivar\cuarto año\primer semestre\sistemas operativos\programas\tarea 1\programa1\programa1\programa1.cpp(54): warning C4101: 'clase1': variable local sin referencia
+c:\users\jcdur\google drive\landivar\cuarto año\primer semestre\sistemas operativos\programas\tarea 1\programa1\programa1\programa1.cpp(58): warning C4101: 'clase2': variable local sin referencia
+  Programa1.vcxproj -> C:\Users\jcdur\Google Drive\Landivar\Cuarto año\Primer Semestre\Sistemas Operativos\Programas\Tarea 1\Programa1\Debug\Programa1.exe
