+++ Programa1/Programa1/Debug/Programa1.tlog/Programa1.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
+Debug|Win32|C:\Users\jcdur\Google Drive\Landivar\Cuarto año\Primer Semestre\Sistemas Operativos\Programas\Tarea 1\Programa1\|
