+++ Programa1/Programa1/ReadMe.txt
@@ -1,0 +1,30 @@
+﻿========================================================================
+    APLICACIÓN DE CONSOLA: Programa1 Información general del proyecto
+========================================================================
+
+AppWizard ha creado esta aplicación Programa1.
+
+Este archivo contiene un resumen de lo que encontrará en cada uno de los archivos que constituyen la aplicación Programa1.
+
+
+Programa1.vcxproj
+    Éste es el archivo de proyecto principal para los proyectos de VC++ generados mediante un Asistente para aplicaciones. Contiene información acerca de la versión de Visual C++ con la que se generó el archivo, así como información acerca de las plataformas, configuraciones y características del proyecto seleccionadas en el Asistente para aplicaciones.
+
+Programa1.vcxproj.filters
+    Éste es el archivo de filtros para los proyectos de VC++ generados mediante un asistente para aplicaciones. Contiene información acerca de la asociación entre los archivos del proyecto y los filtros. Esta asociación se usa en el IDE para mostrar la agrupación de archivos con extensiones similares bajo un nodo específico (por ejemplo, los archivos ".cpp" se asocian con el filtro"Archivos de código fuente").
+
+Programa1.cpp
+    Este es el archivo de código fuente de la aplicación principal.
+
+/////////////////////////////////////////////////////////////////////////////
+Otros archivos estándar:
+
+StdAfx.h, StdAfx.cpp
+    Estos archivos se usan para compilar un archivo de encabezado precompilado (PCH) denominado Programa1.pch y un archivo de tipos precompilados llamado StdAfx.obj.
+
+/////////////////////////////////////////////////////////////////////////////
+Otras notas:
+
+El asistente para aplicaciones utiliza comentarios "TODO:" para indicar las partes del código fuente que tendrá que agregar o personalizar.
+
+/////////////////////////////////////////////////////////////////////////////
