+++ Programa1/Programa1/stdafx.h
@@ -1,0 +1,15 @@
+// stdafx.h: archivo de inclusión de los archivos de inclusión estándar del sistema
+// o archivos de inclusión específicos de un proyecto utilizados frecuentemente,
+// pero rara vez modificados
+//
+
+#pragma once
+
+#include "targetver.h"
+
+#include <stdio.h>
+#include <tchar.h>
+
+
+
+// TODO: mencionar aquí los encabezados adicionales que el programa necesita
