+++ Programa1/Programa1/targetver.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// La inclusión de SDKDDKVer.h define la plataforma Windows más alta disponible.
+
+// Si desea compilar la aplicación para una plataforma Windows anterior, incluya WinSDKVer.h y
+// establezca la macro _WIN32_WINNT en la plataforma que desea admitir antes de incluir SDKDDKVer.h.
+
+#include <SDKDDKVer.h>
