--- Proyecto 1/Programa1/Programa1.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Programa1", "Programa1\Programa1.vcxproj", "{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Debug|x64.ActiveCfg = Debug|x64
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Debug|x64.Build.0 = Debug|x64
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Debug|x86.ActiveCfg = Debug|Win32
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Debug|x86.Build.0 = Debug|Win32
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Release|x64.ActiveCfg = Release|x64
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Release|x64.Build.0 = Release|x64
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Release|x86.ActiveCfg = Release|Win32
-		{FEABFFD9-08AD-4ACD-94AD-4F4A4938C6E1}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
