--- Proyecto 1/Programa1/Programa1/stdafx.cpp
@@ -1,8 +1,0 @@
-// stdafx.cpp: archivo de código fuente que contiene sólo las inclusiones estándar
-// Programa1.pch será el encabezado precompilado
-// stdafx.obj contiene la información de tipos precompilada
-
-#include "stdafx.h"
-
-// TODO: mencionar los encabezados adicionales que se necesitan en STDAFX.H
-// pero no en este archivo
