--- Proyecto 1/README.md
@@ -1,19 +1,0 @@
-#Callbacks
-
-Utilizar Visual C++
-Crear tres funciones bajo las siguientes instrucciones basado en un proyecto de consola.
-
-Todas las Funciones tienen la siguiente firma
-int function(int param);
-
- 
-
-1. La primera función debe definirse como función estática de todo el programa
-
-2. La segunda función debe definirse como función estática de una clase A
-
-3. La tercera función debe de definirse como función privada de una clase B y requiere que el objeto esté inicializado
-
-
-Una vez definidas las funciones, estas deben retornar el # al que corresponde e imprimir en pantalla el parámetro recibido y el numeral.
-Crear tres callbacks que tengan la referencia hacia dada función y deben de ser llamadas desde el main del programa.
